n1sdp: upcast (UINT16_MAX*FWK_MHZ) to uint64_t for comparison
In n1sdp_pll, there is an assert call to check the incoming
rate which is uint64_t with UINT16_MAX. The calculation
fails in debug mode build. This commit fixes the calculation.

+++ product/n1sdp/module/n1sdp_pll/src/mod_n1sdp_pll.c
@@ -68,7 +68,7 @@ static int pll_set_rate(struct n1sdp_pll_dev_ctx *ctx, uint64_t rate,
     size_t i;
 
     fwk_assert(ctx != NULL);
-    fwk_assert(rate <= (UINT16_MAX * FWK_MHZ));
+    fwk_assert(rate <= ((uint64_t)UINT16_MAX * FWK_MHZ));
 
     config = ctx->config;
